Remove mypy usage and add size util
