+++ .pre-commit-config.yaml
@@ -65,12 +65,6 @@ repos:
         name: Ruff check
         args: ["--fix"]
 
-  # it needs to be after formatting hooks because the lines might be changed
-  - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.16.1
-    hooks:
-      - id: mypy
-        files: "src/*"
 
   - repo: https://github.com/tox-dev/pyproject-fmt
     rev: v2.6.0
