+++ pyproject.toml
@@ -46,6 +46,7 @@ dynamic = [
 dependencies = [
   "portalocker>=2.3.2",
   "watchdog>=2.3.1",
+  "pympler>=1.0",
 ]
 urls.Source = "https://github.com/python-cachier/cachier"
 # --- setuptools ---
@@ -197,5 +198,5 @@ show_missing = true
 exclude_lines = [
   "pragma: no cover",          # Have to re-enable the standard pragma
   "raise NotImplementedError", # Don't complain if tests don't hit defensive assertion code:
-  "if TYPE_CHECKING:",         # Is only true when running mypy, not tests
+  "if TYPE_CHECKING:",
 ]
