+++ src/cachier/__init__.py
@@ -4,11 +4,11 @@ from .config import (
     enable_caching,
     get_default_params,
     get_global_params,
-    parse_bytes,
     set_default_params,
     set_global_params,
 )
 from .core import cachier
+from .util import parse_bytes
 
 __all__ = [
     "cachier",
