+++ src/cachier/config.py
@@ -1,7 +1,6 @@
 import hashlib
 import os
 import pickle
-import re
 import threading
 from dataclasses import dataclass, field, replace
 from datetime import datetime, timedelta
@@ -31,27 +30,6 @@ def _default_cache_dir():
         return os.path.join(xdg_cache_home, "cachier")
     # Default fallback if XDG is not set
     return os.path.expanduser("~/.cachier/")
-
-
-def parse_bytes(size: Union[int, str, None]) -> Optional[int]:
-    """Convert a human friendly size string to bytes."""
-    if size is None:
-        return None
-    if isinstance(size, int):
-        return size
-    match = re.fullmatch(r"(?i)\s*(\d+(?:\.\d+)?)\s*([kmgt]?b)?\s*", str(size))
-    if not match:
-        raise ValueError(f"Invalid size value: {size}")
-    number = float(match.group(1))
-    unit = (match.group(2) or "b").upper()
-    factor = {
-        "B": 1,
-        "KB": 1024,
-        "MB": 1024**2,
-        "GB": 1024**3,
-        "TB": 1024**4,
-    }[unit]
-    return int(number * factor)
 
 
 class LazyCacheDir:
